*: refine tests.
