+++ statistics/handle_test.go
@@ -24,6 +24,7 @@ import (
 	"github.com/pingcap/tidb/model"
 	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/statistics"
+	"github.com/pingcap/tidb/store/tikv"
 	"github.com/pingcap/tidb/util/testkit"
 	"github.com/pingcap/tidb/util/types"
 )
@@ -316,10 +317,12 @@ func (s *testStatsCacheSuite) TestLoadHist(c *C) {
 }
 
 func newStoreWithBootstrap() (kv.Storage, *domain.Domain, error) {
-	store, err := tidb.NewStore(tidb.EngineGoLevelDBMemory)
+	store, err := tikv.NewMockTikvStore()
 	if err != nil {
 		return nil, nil, errors.Trace(err)
 	}
+	tidb.SetSchemaLease(0)
+	tidb.SetStatsLease(0)
 	do, err := tidb.BootstrapSession(store)
 	return store, do, errors.Trace(err)
 }
